Handle Error Reading on the Second Day after Daylight Saving Change in the Spring

This internal function removes all rows prior to a time discontinuity caused by a daylight saving time adjustment. Since spring daylight saving change will push the time forward by 1 hour, this causes "time overspill". Ususally on the second day after daylight saving change in the spring, the datasheet starts with before midnight recordings from the previous day, instead of recordings of the current day.

Usage

next_day_after_spring_dst(data_frame, start_col = start)

Arguments

data_frame

A data frame with a time column (formatted as "HH:MM:SS") representing visit start times.

start_col

Name of the time column (unquoted), usually this is the start time of the event. Default is: start.

Value

A filtered data frame with only rows after the DST-induced reset point.

Details

This function detects the first instance where the start time goes backward (i.e., the next row's time is earlier than the previous row's time), and removes all rows before that point.