Process a batch of feeder files

Reads, cleans, and time‑adjusts a batch of files, returning a named list of data frames keyed by file date (YYYY‑MM‑DD).

Usage

process_all_feed(
  files,
  col_names,
  id_col = "cow",
  drop_ids = NULL,
  trans_col = "transponder",
  drop_trans = NULL,
  bin_col = "bin",
  bins,
  select_cols = NULL,
  sep = ",",
  header = FALSE,
  daylight_change_duration = 60,
  tz = Sys.timezone()
)

Arguments

files

Character vector of all the file paths.

col_names

Character vector of names to assign to the raw data columns.

id_col

Single string: column name for animal IDs (default: "cow").

drop_ids

Vector: values in id_col to remove (default: NULL).

trans_col

Single string: column name for transponder IDs (default: "transponder").

drop_trans

Vector: values in trans_col to remove (default: NULL).

bin_col

Single string: column name for bin IDs (default: "bin").

bins

Numeric vector: bin IDs to keep (e.g. 2:4 or c(1,5)).

select_cols

Character vector: columns to retain in the final output. Default is null, select all columns.

sep

Field separator; passed to read.table(). Defaults to ",".

header

Logical; does the file have a header row? Defaults to FALSE.

daylight_change_duration

An integer for the fallback duration in minutes (default = 60).

tz

A valid time zone name (default is the time zone of your current physical location), used to determine DST rules. Use OlsonNames() to see all valid options.

Value

A named list of data frames, one per input file, named by date (YYYY‑MM‑DD).

Details

Steps:

Validate all inputs (types, file_type ∈ "feed","water").

Extract a date information from each filename and parse it via lubridate::ymd().

Fetch Daylight Saving Time (DST) switch table for the relevant years (dst_switch_day(years, tz)).

Loop over each file: 4.1. Call either process_feeder() or process_water() to do: 4.1.1. Safely read the CSV / DAT file 4.1.2. Rename columns 4.1.3. Drop unwanted cows & transponders 4.1.4. Keep only specified bins 4.1.5. Subset to desired columns 4.2. Drop any rows with NA. 4.3. Call daylight_saving_adjust() to adjust timestamps for daylight saving change days. 4.4. Standardize the columns recording start and end time of each event to be in the format of "yyyy-mm-dd hh:mm:ss". 4.5. Store processed dataframe the output list and name it by the date.

See also

Examples

if (FALSE) { # \dontrun{
res <- process_all_feed(
  files       = c("path/to/file/VR210102.DAT", "path/to/file/VW210605.DAT"),
  col_names   = c("cow", "transponder", "bin", "start", "end"),
  bins        = 1:30,
  select_cols = c("cow", "bin", "start", "end"),
  sep         = ",",
  header      = TRUE,
  tz          = "UTC"
)
} # }