Run full quality-control check on feeder and drinker data

Run full quality-control check on feeder and drinker data

Usage

qc(
  feed = NULL,
  water = NULL,
  cfg = qc_config(),
  id_col = id_col2(),
  start_col = start_col2(),
  end_col = end_col2(),
  bin_col = bin_col2(),
  dur_col = duration_col2(),
  intake_col = intake_col2(),
  start_weight_col = start_weight_col2(),
  end_weight_col = end_weight_col2(),
  tz = tz2(),
  verbose = TRUE
)

Arguments

feed

A list of daily feed data frames named by date, or NULL if you don't have feeder data

water

A list of daily water data frames named by date, or NULL if you don't have water data

cfg

A configuration list created by qc_config().

id_col

Animal ID column name (default current global value)

start_col

Start time column name (default current global value)

end_col

End time column name (default current global value)

bin_col

Bin ID column name (default current global value)

dur_col

Duration column name (default current global value)

intake_col

Intake column name (default current global value)

start_weight_col

Start weight column name (default current global value)

end_weight_col

End weight column name (default current global value)

tz

Time zone string for date-time operations

verbose

Logical. If TRUE, print details of data where errors were detected

Value

A list with four elements:

warnings

a tidy data frame with one row per day and one column per warning code.

feed

cleaned feed list (or NULL).

water

cleaned water list (or NULL).

combined

merged feed + water list (after all fixes).

Examples

cfg <- qc_config(high_dur_feed = 2500, low_visit_threshold = 5, total_cows_expected=48)
out <- qc(feed = all_fed, water = all_wat, cfg = cfg, verbose = FALSE)
out$warnings
#> # A tibble: 3 × 17
#>   date       total_cows missing_cow double_detection_bins negative_visit_bins
#>   <chr>           <int> <chr>       <chr>                 <chr>              
#> 1 2020-10-31         47 Yes         30                    6; 12; 13          
#> 2 2020-11-01         47 Yes         3; 5; 6; 18           6; 12; 13; 30      
#> 3 2020-11-02         48 NA          NA                    13; 30             
#> # ℹ 12 more variables: cows_disappeared_after_noon <chr>,
#> #   bins_never_visited <chr>, bins_low_traffic <chr>, long_dur_feeder <chr>,
#> #   large_intake_feed_visit <chr>, low_daily_feed_intake_cows <chr>,
#> #   high_daily_feed_intake_cows <chr>, feed_add_time_no_found <chr>,
#> #   long_dur_drinker <chr>, large_intake_water_visit <chr>,
#> #   low_daily_water_intake_cows <chr>, high_daily_water_intake_cows <chr>