Replacement events between cows

A dataset containing validated replacement events where one cow (actor) replaces another cow (reactor) at a feeding bin across the three-day period (2020-10-31 to 2020-11-02). This dataset is the result of applying the record_replacement_days() function to the quality-checked feeding data.

Usage

replacements

Format

A list of 3 data frames, one for each date (2020-10-31, 2020-11-01, 2020-11-02), with each data frame containing the following variables:

reactor_cow

integer, ID of the cow that was replaced (had to leave the bin)

bin

integer, feeding bin location number where the replacement occurred

time

POSIXct, timestamp when the replacement event occurred

date

Date, calendar date of the replacement event

actor_cow

integer, ID of the cow that initiated the replacement (took over the bin)

bout_interval

difftime, time gap in seconds between when the reactor cow left and the actor cow arrived

Source

Generated by applying record_replacement_days() to clean_feed

Details

A replacement event is defined as an instance where one cow (reactor) leaves a bin and another cow (actor) enters the same bin within a short time threshold (default: 26 seconds). This dataset only includes validated replacements where the actor cow did not have an "alibi" (i.e., was not recorded at another bin at the same time), increasing confidence that these represent true social displacement events.

Examples

# Access replacement events for the first day
first_day_replacements <- replacements[[1]]

# Count the number of times each cow was replaced (as reactor)
reactor_counts <- table(first_day_replacements$reactor_cow)
head(sort(reactor_counts, decreasing = TRUE))
#> 
#> 7030 7027 7018 6090 6028 6084 
#>   29   25   23   21   20   20 

# Count the number of times each cow replaced others (as actor)
actor_counts <- table(first_day_replacements$actor_cow)
head(sort(actor_counts, decreasing = TRUE))
#> 
#> 6050 5120 5058 7027 7018 5124 
#>   34   29   27   25   24   23