Set multiple global variables at once

This function allows users to set multiple global variables simultaneously. Each parameter defaults to its current global value if unspecified.

Usage

set_global_cols(
  tz = tz2(),
  id_col = id_col2(),
  trans_col = trans_col2(),
  start_col = start_col2(),
  end_col = end_col2(),
  bin_col = bin_col2(),
  dur_col = duration_col2(),
  intake_col = intake_col2(),
  start_weight_col = start_weight_col2(),
  end_weight_col = end_weight_col2(),
  bin_offset = bin_offset2(),
  bins_feed = bins_feed2(),
  bins_wat = bins_wat2(),
  bin_layout = bin_layout2()
)

Arguments

tz

Timezone (default current global value from tz2())

id_col

Animal ID column name (default current global value from id_col2())

trans_col

Transponder column name (default current global value from trans_col2())

start_col

Start time column name (default current global value from start_col2())

end_col

End time column name (default current global value from end_col2())

bin_col

Bin ID column name (default current global value from bin_col2())

dur_col

Duration column name (default current global value from duration_col2())

intake_col

Intake column name (default current global value from intake_col2())

start_weight_col

Start weight column name (default current global value from start_weight_col2())

end_weight_col

End weight column name (default current global value from end_weight_col2())

bin_offset

Numeric bin offset (default current global value from bin_offset2())

bins_feed

Integer vector of feed bins (default current global value from bins_feed2())

bins_wat

Integer vector of water bins (default current global value from bins_wat2())

bin_layout

Character string of physical bin layout with rows separated by \n and bins within rows separated by "-" (default current global value from bin_layout2())

Value

Called for its side-effects

Examples

set_global_cols(tz = "UTC", id_col = "animal_id", dur_col = "visit_duration")
#> Note: Bin offset is set to 50. During data processing, water bins 1-3 will be transformed to 51-53.
#> Note: The following bins from your feed/water bin lists are not included in the layout. Please ignore this message if this is intended. Otherwise you need to update your bin layout: 15, 16, 17, 18, 19, 20, 51, 52, 53